requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
click==6.7
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.9
cycler==0.10.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-celery==3.2.1
django-colorful==1.2
django-filter==1.0.4
django-leaflet==0.22.0
django-raster==0.3.1
django-toolbelt==0.0.1
djangorestframework==3.6.3
djangorestframework-gis==0.11.2
enum34==1.1.6
Fiona==1.7.8
Flask==0.12.1
flufl.enum==4.1.1
freeze==1.0.10
gdal==2.1.3
gunicorn==19.7.1
heroku==0.1.4
html5lib==0.999999999
httpie==0.9.9
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
kombu==3.0.37
lxml==3.7.3
MarkupSafe==1.0
matplotlib==2.0.0
munch==2.1.1
numpy==1.12.1
olefile==0.44
owslib==0.14.0
pbr==2.0.0
Pillow==4.1.0
psycopg2==2.7.1
Pygments==2.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyproj==1.9.5.1
pyshp==1.2.10
python-dateutil==1.5
pytz==2017.2
redis==2.10.5
reportlab==3.4.0
requests==2.13.0
six==1.10.0
sqlalchemy==1.1.9
static3==0.7.0
stevedore==1.21.0
vine==1.1.3
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
webencodings==0.5.1
werkzeug==0.12.1
whitenoise==3.3.0
xhtml2pdf==0.0.6
